requirements.txt
# requirements.txt

# Основные библиотеки
python-telegram-bot==20.7
python-dotenv==1.0.0
Flask==3.0.0
requests==2.31.0
Flask-SocketIO==5.3.6

# Утилиты для работы с датами и временем
pytz==2023.3

# Логирование
colorlog==6.7.0